[iOS] chore: Usecase 그룹화
